requirements.txt
fastapi
uvicorn
openai
pymongo
python-dotenv
tiktoken
